+++ elements/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `@pod-os/elements` now includes a [Custom Elements Manifest](https://github.com/webcomponents/custom-elements-manifest) at `dist/custom-elements.json` that can be used in IDEs to
   enable auto-completion for web components.
 - [pos-add-relation](https://pod-os.org/reference/elements/components/pos-add-relation/): New component to add a relation (link) to a thing
+- [pos-relations](https://pod-os.org/reference/elements/components/pos-relations/): Allows to add a new relation, if resource is editable
 
 ### Fixed